Fix crash for cobertura coverage converter
The DTD URL changed and the fallback file, which was supposed to be read from the bundle crashed the app.
No surprise, because the command line tool is not a bundle. DUH!
The fix is to include the DTD as string in the code. Also removed the online poll of the DTD, as it will probably never change!