Add all failures for a given test to the JUnit report
Previously, we were only ever parsing the first failure for a test and
passing that to the JUnit representation. However, a test might contain
multiple failures and this is fully supported by the JUnit spec.

To fix the issue, we'll collect all failures from the XCResult file and
pass them to the JUnit result a list. This will prevent us from losing
important diagnostic information from the test result.